Resolve 64-bit time_t issues with combination MSVC10/BCB6 builds by making all exported functions and data types using a 32-bit time_t (time32_t).

+++ src/sbbs3/inkey.cpp
@@ -238,7 +238,7 @@ char sbbs_t::handle_ctrlkey(char ch, long mode)
 			bprintf(text[TiLogon],timestr(logontime));
 			bprintf(text[TiNow],timestr(now));
 			bprintf(text[TiTimeon]
-				,sectostr(now-logontime,tmp));
+				,sectostr((uint)(now-logontime),tmp));
 			bprintf(text[TiTimeLeft]
 				,sectostr(timeleft,tmp));
 			SYNC;
